+++ src/shared/components/ui/Skeleton.test.tsx
@@ -4,7 +4,7 @@
  */
 
 import React from 'react';
-import { render, screen } from '../../../test-utils';
+import { render } from '../../../test-utils';
 import {
   Skeleton,
   SkeletonText,